refactor: move upload functions to lib folder
