+++ frontend/src/app/http/new/page.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import { SubmitButton } from "@/components/general/general";
-import { FileInputRow, ShortInputRow } from "@/components/general/inputs";
+import { ImageFileInputRow, ShortInputRow } from "@/components/general/inputs";
+import { uploadMedia } from "@/lib/upload";
 import axios from "axios";
 import { useForm } from "react-hook-form";
 
@@ -13,35 +14,11 @@ export interface IHttpInputForm {
 }
 
 interface IHttpRequest extends IHttpInputForm {
-  imagePath: string;
+  imagePath?: string;
 }
 
 const HttpAddPage: React.FC = () => {
   const { register, handleSubmit, reset } = useForm<IHttpInputForm>();
-
-  const uploadImage = async (data: IHttpInputForm) => {
-    const formData = new FormData();
-    if (data.image && data.image.length > 0) {
-      formData.append("image", data.image[0]);
-      formData.append("imageId", `http_${data.code}_${data.image[0].name}`);
-    }
-
-    try {
-      const response = await axios.post("/api/v0/upload", formData, {
-        headers: {
-          "Content-Type": "multipart/form-data"
-        }
-      });
-
-      if (response.data && response.data.imagePath) {
-        return response.data.imagePath;
-      }
-      return "";
-    } catch (error) {
-      console.error(`${error} occurred`);
-      return "";
-    }
-  };
 
   const uploadHttp = async (data: IHttpRequest) => {
     try {
@@ -68,11 +45,10 @@ const HttpAddPage: React.FC = () => {
       return;
     }
 
-    // Null Check
-    const imagePath = await uploadImage(data);
+    const imagePath = await uploadMedia(data, "image");
     const updatedData = {
       ...data,
-      imagePath
+      ...(imagePath && { imagePath })
     };
 
     await uploadHttp(updatedData);
@@ -83,7 +59,7 @@ const HttpAddPage: React.FC = () => {
         <ShortInputRow register={register("code")} label="코드" />
         <ShortInputRow register={register("message")} label="메시지" />
         <ShortInputRow register={register("description")} label="설명" />
-        <FileInputRow register={register("image")} label="이미지" />
+        <ImageFileInputRow register={register("image")} label="이미지" />
         <SubmitButton />
       </div>
     </form>
