+++ frontend/src/app/quiz/new/page.tsx
@@ -8,6 +8,7 @@ import {
   SelectInputRow,
   ShortInputRow
 } from "@/components/general/inputs";
+import { uploadFile, uploadMedia } from "@/lib/upload";
 import axios from "axios";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
@@ -29,60 +30,6 @@ interface IQuizRequest extends IQuizInputForm {
   audioPath?: string;
   audiosPath?: string[];
 }
-
-const uploadApi = async (
-  file: File,
-  fileName: string,
-  type: "image" | "audio"
-): Promise<string> => {
-  const formData = new FormData();
-
-  formData.append(type, file);
-  formData.append(`${type}Id`, fileName);
-
-  try {
-    const { data } = await axios.post(`/api/v0/upload/${type}`, formData, {
-      headers: {
-        "Content-Type": "multipart/form-data"
-      }
-    });
-
-    return data?.filePath || "";
-  } catch (error) {
-    console.error(error);
-    return "";
-  }
-};
-
-const uploadFile = async (
-  file: File,
-  data: IQuizRequest,
-  type: "image" | "audio",
-  index?: number
-): Promise<string> => {
-  const suffix = getSuffix(file.name);
-  const fn =
-    index === undefined
-      ? `quiz_${data.quizSetId}_${data.problemId}_${type}.${suffix}`
-      : `quiz_${data.quizSetId}_${data.problemId}_${type}${index}.${suffix}`;
-
-  return await uploadApi(file, fn, type);
-};
-
-const getSuffix = (filename: string): string => {
-  const suffix = filename.split(".").pop();
-  return suffix || "";
-};
-
-const uploadMedia = async (
-  data: IQuizRequest,
-  type: "image" | "audio"
-): Promise<string> => {
-  const file = data[type]?.[0];
-  if (!file) return "";
-
-  return await uploadFile(file, data, type);
-};
 
 const uploadAudios = async (data: IQuizRequest): Promise<string[]> => {
   if (!data.audios) {
